Began writing readIn() for SceneLayout
Wrote Block class so that it could be used in readIn() and SceneLayout.
Altered the packages contained within /media/, images are now stored in
/media/images/, levels are stored in /media/levels/. Moved files
accordingly.

NOTE: Block and readIn() are not fully functional as of this commit.
Block's constructor is being worked on, part of which requires
determining what sprites are going to be used for Afterdark and
enumerating them so that the constuctor for Block can easily assign
images.

readIn() requires me reading a little bit more on how .csv files are
best read in Java. At the moment, I'm reading it like a normal text file
and the parsing through despite there probably being a better method for
doing this.